Feat: replyAPI api 게시글, 댓글, 답글 존재하는지 확인 로직 및 예외 처리 로직 추가, fetchReportReply 함수 userDoc에 reportReplyList 추가 로직 삭제 답글 첫 페이지 조회 fetchFirstPageReplyData 함수 댓글 존재 확인 및 예외 처리 추가 답글 추가 fetchAddReply 함수 게시물, 댓글 존재 확인 및 예외 처리 추가 답글 수정 fetchEditReply 함수 게시물, 댓글, 답글 존재 확인 및 예외 처리 추가 답글 삭제 fetchRemoveReply 함수 게시물, 댓글, 답글 존재 확인 및 예외 처리 추가 userDoc에 reportReplyList 추가 로직 삭제 대신 replyDoc 자체에 reportUidList 추가하는 로직 추가

+++ src/api/firebase/replyAPI.ts
@@ -13,7 +13,8 @@ import {
   QueryDocumentSnapshot,
   DocumentData,
   startAfter,
-  where
+  where,
+  getDoc
 } from "firebase/firestore";
 import { IReplyData, IUserData } from "../apiType";
 import { db } from "./setting";
@@ -29,7 +30,13 @@ export const fetchFirstPageReplyData = async (
   pagePerData: number
 ) => {
   try {
+    // 댓글 확인 및 예외처리
     const commentDoc = doc(db, `comments/${parentCommentId}`);
+    const commentDocSnapShot = await getDoc(commentDoc);
+    if(!commentDocSnapShot.exists()){
+      throw new Error("댓글이 존재하지 않습니다.")
+    }
+
     const replyRef = collection(commentDoc, "replies");
     const q = query(replyRef, orderBy("createdAt", "desc"), limit(pagePerData));
     const replyDoc = await getDocs(q);
@@ -113,7 +120,29 @@ export const fetchPagingReplyData = async (
  */
 export const fetchAddReply = async (replyCommentData: IReplyData) => {
   try {
+    // 병렬 처리할 비동기 작업 배열 생성
+    const tasks = [];
+
+    // 게시물 확인 작업 추가
+    const postDoc = doc(db, `post/${replyCommentData.postId}`);
+    tasks.push(getDoc(postDoc));
+
+    // 댓글 확인 작업 추가
     const commentDoc = doc(db, `comments/${replyCommentData.parentCommentId}`);
+    tasks.push(getDoc(commentDoc));
+
+    // 모든 작업을 병렬로 실행하고 결과 배열을 얻습니다.
+    const results = await Promise.all(tasks);
+
+    // 결과 확인 및 예외 처리
+    if (!results[0].exists()) {
+      throw new Error("게시물이 존재하지 않습니다.");
+    }
+
+    if (!results[1].exists()) {
+      throw new Error("댓글이 존재하지 않습니다.");
+    }
+
     const replyRef = collection(commentDoc, "replies");
     // 댓글 데이터 subCollection replies에 답글 Doc 추가
     const addReplyCommentProimse = setDoc(
@@ -137,11 +166,43 @@ export const fetchAddReply = async (replyCommentData: IReplyData) => {
  * 답글 수정
  */
 export const fetchEditReply = async (
-  replyEditData: Pick<IReplyData, "parentCommentId" | "replyId" | "content">
+  replyEditData: Pick<
+    IReplyData,
+    "parentCommentId" | "replyId" | "content" | "postId"
+  >
 ) => {
   try {
+    // 병렬 처리할 비동기 작업 배열 생성
+    const tasks = [];
+
+    // 게시물 확인 작업 추가
+    const postDoc = doc(db, `post/${replyEditData.postId}`);
+    tasks.push(getDoc(postDoc));
+
+    // 댓글 확인 작업 추가
     const commentDoc = doc(db, `comments/${replyEditData.parentCommentId}`);
+    tasks.push(getDoc(commentDoc));
+
+    // 답글 확인 작업 추가
     const replyDoc = doc(commentDoc, `replies/${replyEditData.replyId}`);
+    tasks.push(getDoc(replyDoc));
+
+    // 모든 작업을 병렬로 실행하고 결과 배열을 얻습니다.
+    const results = await Promise.all(tasks);
+
+    // 결과 확인 및 예외 처리
+    if (!results[0].exists()) {
+      throw new Error("게시물이 존재하지 않습니다.");
+    }
+
+    if (!results[1].exists()) {
+      throw new Error("댓글이 존재하지 않습니다.");
+    }
+
+    if (!results[2].exists()) {
+      throw new Error("답글이 존재하지 않습니다.");
+    }
+
     // 답글 doc의 content 수정
     await updateDoc(replyDoc, {
       content: replyEditData.content
@@ -156,11 +217,40 @@ export const fetchEditReply = async (
  * 답글 삭제
  */
 export const fetchRemoveReply = async (
-  replyEditData: Pick<IReplyData, "parentCommentId" | "replyId">
+  replyRemoveData: Pick<IReplyData, "parentCommentId" | "replyId" | "postId">
 ) => {
   try {
-    const commentDoc = doc(db, `comments/${replyEditData.parentCommentId}`);
-    const replyDoc = doc(commentDoc, `replies/${replyEditData.replyId}`);
+    // 병렬 처리할 비동기 작업 배열 생성
+    const tasks = [];
+
+    // 게시물 확인 작업 추가
+    const postDoc = doc(db, `post/${replyRemoveData.postId}`);
+    tasks.push(getDoc(postDoc));
+
+    // 댓글 확인 작업 추가
+    const commentDoc = doc(db, `comments/${replyRemoveData.parentCommentId}`);
+    tasks.push(getDoc(commentDoc));
+
+    // 답글 확인 작업 추가
+    const replyDoc = doc(commentDoc, `replies/${replyRemoveData.replyId}`);
+    tasks.push(getDoc(replyDoc));
+
+    // 모든 작업을 병렬로 실행하고 결과 배열을 얻습니다.
+    const results = await Promise.all(tasks);
+
+    // 결과 확인 및 예외 처리
+    if (!results[0].exists()) {
+      throw new Error("게시물이 존재하지 않습니다.");
+    }
+
+    if (!results[1].exists()) {
+      throw new Error("댓글이 존재하지 않습니다.");
+    }
+
+    if (!results[2].exists()) {
+      throw new Error("답글이 존재하지 않습니다.");
+    }
+
     // 답글 doc 삭제
     const removeReplyPromise = deleteDoc(replyDoc);
 
@@ -182,15 +272,50 @@ export const fetchRemoveReply = async (
 export const fetchReportReply = async (
   replyReportData: Pick<
     IReplyData,
-    "replyId" | "parentCommentId" | "reportCount"
+    "replyId" | "parentCommentId" | "reportCount" | "postId"
   >
 ) => {
   try {
+    // 병렬 처리할 비동기 작업 배열 생성
+    const tasks = [];
+
+    // 게시물 확인 작업 추가
+    const postDoc = doc(db, `post/${replyReportData.postId}`);
+    tasks.push(getDoc(postDoc));
+
+    // 댓글 확인 작업 추가
     const commentDoc = doc(db, `comments/${replyReportData.parentCommentId}`);
+    tasks.push(getDoc(commentDoc));
+
+    // 답글 확인 작업 추가
     const replyDoc = doc(commentDoc, `replies/${replyReportData.replyId}`);
+    tasks.push(getDoc(replyDoc));
+
+    // 모든 작업을 병렬로 실행하고 결과 배열을 얻습니다.
+    const results = await Promise.all(tasks);
+
+    // 결과 확인 및 예외 처리
+    if (!results[0].exists()) {
+      throw new Error("게시물이 존재하지 않습니다.");
+    }
+
+    if (!results[1].exists()) {
+      throw new Error("댓글이 존재하지 않습니다.");
+    }
+
+    if (!results[2].exists()) {
+      throw new Error("답글이 존재하지 않습니다.");
+    }
+
+    const replySnapShot = await getDoc(replyDoc);
+    if (!replySnapShot.exists()) {
+      throw new Error("답글이 존재하지 않습니다.");
+    }
+    if(!auth.currentUser) return;
     const addReportCountPromise = await updateDoc(replyDoc, {
       reportCount: increment(1),
-      isBlock: replyReportData.reportCount >= 4 ? true : false
+      isBlock: replyReportData.reportCount >= 4 ? true : false,
+      reportUidList: arrayUnion(auth.currentUser.uid)
     });
 
     const decreaseReplyCountPromise = [];
@@ -201,15 +326,9 @@ export const fetchReportReply = async (
           replyCount: increment(-1)
         })
       );
-    }
+      }
 
-    if (!auth.currentUser) return;
-    const userDoc = doc(db, `user/${auth.currentUser.uid}`);
-    const addReportListPromise = updateDoc(userDoc, {
-      reportReplyList: arrayUnion(replyReportData.replyId)
-    });
-
-    await Promise.all([addReportCountPromise, addReportListPromise]);
+    await Promise.all([addReportCountPromise, ...decreaseReplyCountPromise]);
   } catch (error) {
     console.error(error);
     throw error;
